requirements.txt
streamlit==1.51.0
pandas==2.3.3
langchain-openai==0.3.35
langchain-experimental==0.3.4
langgraph==0.2.74
python-dotenv==1.2.1
